+++ src/renderer/Renderer.h
@@ -34,9 +34,9 @@ protected:
     int m_renderFrameStart = 0;
     int m_renderFrameStep = 0;
 
-    float m_elapsedSecs = -1;
+    double m_elapsedSecs = -1;
 
-    float m_lastHitTime = 0.f;
+    double m_lastHitTime = 0.f;
     unsigned long m_lastHitFrame = 0;
 
     bool m_rendered = false;
